requirements.txt
aerich
fastapi
tortoise-orm
uvicorn[standard]
strawberry-graphql[fastapi]